# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Event",
        "version": "1.0.0"
    },
    "paths": {
        "/api/events/{eventId}/reserve": {
            "post": {
                "tags": [
                    "api:ticket:events"
                ],
                "summary": "Забронировать места события",
                "operationId": "8bde083d27602babfc26876e05fafbe1",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "параметры",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "places[]"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "places[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventBookSeatsResponseDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "EventBookSeatsResponseDTO": {
                "title": "response",
                "description": "Бронирование",
                "properties": {
                    "response": {
                        "$ref": "#/components/schemas/ReserveResponseDTO"
                    }
                },
                "type": "object"
            },
            "ReserveResponseDTO": {
                "title": "reserve",
                "description": "объект бронирование",
                "properties": {
                    "reservation_id": {
                        "description": "ID резерва",
                        "type": "string"
                    },
                    "success": {
                        "description": "Результат бронирования",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}